+++ R/pas_downloadParseRawData.R
@@ -41,7 +41,7 @@
 #'
 #' pas_raw <-
 #'   pas_downloadParseRawData(
-#'     apiReadKey = API_READ_KEY,
+#'     apiReadKey = PA_API_READ_KEY,
 #'     maxAge = 3600 * 24,
 #'     outsideOnly = TRUE,
 #'     west = -125,
@@ -184,7 +184,8 @@ pas_downloadParseRawData <- function(
       "web service error %s from:\n  %s\n\n%s",
       status_code,
       webserviceUrl,
-      httpcode::http_code(status_code)$explanation
+      ###httpcode::http_code(status_code)$explanation
+      http_status_codes[[as.character(status_code)]][[2]]
     )
 
     if ( logger.isInitialized() ) {